Complete Assignment5 authentication and authorization

+++ src/api/models/cat-model.js
@@ -50,11 +50,20 @@ const addCat = async (cat) => {
   };
 };
 
-const modifyCat = async (cat, id) => {
-  const sql = promisePool.format('UPDATE wsk_cats SET ? WHERE cat_id = ?', [
-    cat,
-    id,
-  ]);
+const modifyCat = async (cat, id, user) => {
+  let sql;
+
+  if (user.role === 'admin') {
+    sql = promisePool.format('UPDATE wsk_cats SET ? WHERE cat_id = ?', [
+      cat,
+      id,
+    ]);
+  } else {
+    sql = promisePool.format(
+      'UPDATE wsk_cats SET ? WHERE cat_id = ? AND owner = ?',
+      [cat, id, user.user_id],
+    );
+  }
 
   const [result] = await promisePool.query(sql);
 
@@ -67,11 +76,20 @@ const modifyCat = async (cat, id) => {
   };
 };
 
-const removeCat = async (id) => {
-  const [result] = await promisePool.execute(
-    'DELETE FROM wsk_cats WHERE cat_id = ?',
-    [id],
-  );
+const removeCat = async (id, user) => {
+  let result;
+
+  if (user.role === 'admin') {
+    [result] = await promisePool.execute(
+      'DELETE FROM wsk_cats WHERE cat_id = ?',
+      [id],
+    );
+  } else {
+    [result] = await promisePool.execute(
+      'DELETE FROM wsk_cats WHERE cat_id = ? AND owner = ?',
+      [id, user.user_id],
+    );
+  }
 
   if (result.affectedRows === 0) {
     return false;
